Keep patching pyOpenSSL runtime for older OpenSSL installs
This reinstates the patch in our pyOpenSSL branch, but at runtime.
That will cause OpenSSL to only emit utf8 strings, rather than mix
and blend with T61 strings that kill Unicode characters.

The defaults has changed in OpenSSL (first in the config file for
users of the command line tool, and then in mainline after we filed
this bug report, but not everyone is on a current OpenSSL)