OpenSSL 1.1 compat hack
This is so ugly that I'm ashamed of myself, but it also sort of works.

OpenSSL changed how it parses Subjects in between 1.0 and 1.1, and now outputs
differently abled utf8-escaped data, and outputs a format that it doesn't
accept as input anymore.

This was clearly a pain, so there's some bruteforce search & replace to attempt
to fix that.

This code is not technically correct, but it works in practice.